chore: split sliderticks and timecontrol
